pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xiaotie"
version = "2.1.0"
description = "小铁 - 轻量级 AI Agent 框架"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,<3.13"
authors = [
    { name = "Leo" }
]
keywords = ["ai", "agent", "llm", "claude", "openai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pydantic>=2.0",
    "tiktoken>=0.7.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "prompt_toolkit>=3.0",
    "aiosqlite>=0.19.0",
    "aiofiles>=23.0",
    "psutil>=5.8.0",
    "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
tui = [
    "textual>=0.50.0",
]
search = [
    "chromadb>=0.4.0",
]
proxy = [
    "mitmproxy>=10.0.0,<11.1.0",
]
scraper = [
    "aiohttp>=3.9.0",
    "tqdm>=4.65.0",
]
repomap = [
    "tree-sitter>=0.21.0",
    "tree-sitter-languages>=1.10.0",
    "networkx>=3.0",
]
secrets = [
    "keyring>=25.0",
]
automation = [
    "pyobjc-framework-Quartz>=10.0",
    "Appium-Python-Client>=4.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "textual>=0.50.0",
    "aiohttp>=3.9.0",
    "tqdm>=4.65.0",
    "mitmproxy>=10.0.0,<11.1.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-mermaid2-plugin>=1.0.0",
]
all = [
    "textual>=0.50.0",
    "chromadb>=0.4.0",
    "mitmproxy>=10.0.0,<11.1.0",
    "aiohttp>=3.9.0",
    "tqdm>=4.65.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-languages>=1.10.0",
    "networkx>=3.0",
    "pyobjc-framework-Quartz>=10.0; sys_platform == 'darwin'",
    "Appium-Python-Client>=4.0.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-mermaid2-plugin>=1.0.0",
]

[project.scripts]
xiaotie = "xiaotie.cli:main"
xiaotie-tui = "xiaotie.tui.main:run_tui"

[project.urls]
Homepage = "https://github.com/LeoLin990405/xiaotie"
Documentation = "https://github.com/LeoLin990405/xiaotie#readme"
Repository = "https://github.com/LeoLin990405/xiaotie"

[tool.hatch.build.targets.wheel]
packages = ["xiaotie"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.bandit]
exclude_dirs = ["tests", "benchmarks", "scripts"]
skips = ["B101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=xiaotie --cov-report=term-missing --cov-report=html:htmlcov"
markers = [
    "smoke: 核心业务冒烟测试",
]
